Move guids to util.
