+++ src/CMakeLists.txt
@@ -455,7 +455,6 @@ set(common_SRCS
 	environment.cpp
 	filesys.cpp
 	gettext.cpp
-	guid.cpp
 	inventorymanager.cpp
 	itemdef.cpp
 	light.cpp
