+++ src/util/guid.cpp
