+++ src/util/guid.h
